fix(editor): remove forward-looking toolProviders/isProviderRegistered refs
Strip editor references to core exports that don't exist on main yet:
- resolveStoredToolProviders / ToolProviders from @mastra/core/tool-provider
- UnknownToolProviderError from @mastra/core/tool-provider
- isProviderRegistered from @mastra/core/llm

These will be restored when PR #16934 (v1 ToolProvider extensions) lands.

Also moves snapshotsMatch to an internal module (not re-exported from
package root) to avoid expanding the public API surface.
